backend/requirements.txt
Flask
Flask-CORS
psycopg2-binary
python-dotenv
Flask-SocketIO